swal en eliminar usuario y en lista de usuarios admin

+++ client/src/pages/users/UserProfilePage.jsx
@@ -7,8 +7,9 @@ import { useNavigate } from 'react-router-dom';
 import HackathonList from '../../components/HackathonList.jsx';
 
 import toast from 'react-hot-toast';
+import Swal from 'sweetalert2';
 
-const { VITE_API_UPLOADS, VITE_API_URL } = import.meta.env;
+const { VITE_API_UPLOADS, VITE_API_URL, VITE_APP_NAME } = import.meta.env;
 
 ////////////////
 
@@ -56,26 +57,48 @@ const UserProfilePage = () => {
 
     const handleRemoveUser = async () => {
         try {
-            const userId = authUser.id;
-
-            const res = await fetch(`${VITE_API_URL}/users/delete/${userId}`, {
-                method: 'DELETE',
-                headers: { Authorization: authToken },
+            // Confirmación con sweetalert2.
+            const result = await Swal.fire({
+                title: '¿Estás seguro de que quieres eliminar tu cuenta?',
+                text: 'No podrás revertir esta acción.',
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#FF3333',
+                cancelButtonColor: '#22577A',
+                confirmButtonText: 'Sí, quiero eliminar mi cuenta',
+                cancelButtonText: 'No, me quedo',
             });
-            const body = await res.json();
 
-            if (body.status === 'error') throw new Error(body.message);
-            authLogoutState();
-            toast.success(body.message);
-            navigate('/');
+            // Si el usuario confirma, procedemos con la eliminación.
+            if (result.isConfirmed) {
+                const userId = authUser.id;
+
+                const res = await fetch(
+                    `${VITE_API_URL}/users/delete/${userId}`,
+                    {
+                        method: 'DELETE',
+                        headers: { Authorization: authToken },
+                    },
+                );
+                const body = await res.json();
+
+                if (body.status === 'error') throw new Error(body.message);
+
+                // Mostramos confirmación de eliminación con SweetAlert
+                await Swal.fire({
+                    title: 'Cuenta eliminada',
+                    text: `Gracias por tu paso por ${VITE_APP_NAME}, esperamos verte de nuevo.`,
+                    icon: 'success',
+                });
+
+                authLogoutState();
+                toast.success(body.message);
+                navigate('/');
+            }
         } catch (err) {
             toast.error(err.message, { id: 'userprofile' });
         }
     };
-
-    if (!authUser || !hackathons) {
-        return <div>Cargando...</div>;
-    }
 
     return (
         <div>
@@ -145,7 +168,7 @@ const UserProfilePage = () => {
                 {!historico && <HackathonList hackathons={actualHackathons} />}
             </div>
             <button onClick={handleRemoveUser} className="button-angled-red">
-                Eliminar Usuario
+                Eliminar usuario
             </button>
         </div>
     );
